Relocate `TestCase` to the `ipl\Html\Test` namespace
The class was only reachable via `autoload-dev`, so downstream packages
had no way to extend or use it in their own test suites. Moving it to
`ipl\Html\Test` exposes it through the library's existing autoload
entry, allowing consumers to rely on it without adjusting their
`composer.json`.
