+++ tests/FormDecorator/LabelDecoratorTest.php
@@ -9,7 +9,7 @@ use ipl\Html\FormElement\FieldsetElement;
 use ipl\Html\FormElement\SubmitButtonElement;
 use ipl\Html\FormElement\SubmitElement;
 use ipl\Html\FormElement\TextElement;
-use ipl\Tests\Html\TestCase;
+use ipl\Html\Test\TestCase;
 
 class LabelDecoratorTest extends TestCase
 {
